Add the possibility to customize the uncaught exception handler #178
